Use Cryptodome for RSA signing.
Tor has a custom PKCS#1 padding format (does not include the
algorithmIdentifier as specified in RFC3447), so we had to monkey-patch the
padding function of Cryptodome for this to work...

Because of the change to Cryptodome we also had to change the interface of
sign_digest() to actually take the message and not the digest directly. That's
because Cryptodome wants to hash the message itself.

Co-authored-by: Federico Ceratto <federico.ceratto@gmail.com>